+++ Projects/CoX/Servers/MapServer/MapServer.h
@@ -43,7 +43,7 @@ public:
         MapManager &            map_manager();
         void                    set_game_server_owner(uint8_t owner_id);
         bool                    session_has_xfer_in_progress(uint64_t session_token);
-        uint8_t                 session_map_xfer_idx(uint64_t session_token);
+        uint32_t                 session_map_xfer_idx(uint64_t session_token);
         void                    session_xfer_complete(uint64_t session_token);
 private:
         bool                    Run();
@@ -60,7 +60,7 @@ private:
         uint8_t                         m_owner_game_server_id = INVALID_GAME_SERVER_ID;
         ACE_INET_Addr                   m_base_location; //! this is the base map instance address
         ACE_INET_Addr                   m_base_listen_point; //! this is used as a base map listening endpoint
-        std::map<uint64_t, uint8_t>     m_current_map_xfers;    // Current map transfers in progress on this map server.
+        std::map<uint64_t, uint32_t>     m_current_map_xfers;    // Current map transfers in progress on this map server.
 };
 
 extern MapServer *g_GlobalMapServer;
